refactor in creating and editing done + validation
TODO
- Unit tests
