+++ COMETwebapp.Tests/Components/ReferenceData/CategoriesTableTestFixture.cs
@@ -1,34 +1,30 @@
 // --------------------------------------------------------------------------------------------------------------------
-// <copyright file="CategoriesTableTestFixture.cs" company="Starion Group S.A.">
-//    Copyright (c) 2023-2024 Starion Group S.A.
-//
-//    Authors: Sam Gerené, Alex Vorobiev, Alexander van Delft, Jaime Bernar, Nabil Abbar
-//
-//    This file is part of CDP4-COMET WEB Community Edition
-//    The CDP4-COMET WEB Community Edition is the Starion Web Application implementation of ECSS-E-TM-10-25 Annex A and Annex C.
-//
-//    The CDP4-COMET WEB Community Edition is free software; you can redistribute it and/or
-//    modify it under the terms of the GNU Affero General Public
-//    License as published by the Free Software Foundation; either
-//    version 3 of the License, or (at your option) any later version.
-//
-//    The CDP4-COMET WEB Community Edition is distributed in the hope that it will be useful,
-//    but WITHOUT ANY WARRANTY; without even the implied warranty of
-//    MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the GNU
+//  <copyright file="CategoriesTableTestFixture.cs" company="Starion Group S.A.">
+//     Copyright (c) 2024 Starion Group S.A.
+// 
+//     Authors: Sam Gerené, Alex Vorobiev, Alexander van Delft, Jaime Bernar, Théate Antoine, João Rua
+// 
+//     This file is part of COMET WEB Community Edition
+//     The COMET WEB Community Edition is the Starion Group Web Application implementation of ECSS-E-TM-10-25 Annex A and Annex C.
+// 
+//     The COMET WEB Community Edition is free software; you can redistribute it and/or
+//     modify it under the terms of the GNU Affero General Public
+//     License as published by the Free Software Foundation; either
+//     version 3 of the License, or (at your option) any later version.
+// 
+//     The COMET WEB Community Edition is distributed in the hope that it will be useful,
+//     but WITHOUT ANY WARRANTY; without even the implied warranty of
+//     MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the GNU
 //    Affero General Public License for more details.
-//
+// 
 //    You should have received a copy of the GNU Affero General Public License
 //    along with this program.  If not, see <http://www.gnu.org/licenses/>.
-// </copyright>
-// --------------------------------------------------------------------------------------------------------------------
+//  </copyright>
+//  --------------------------------------------------------------------------------------------------------------------
 
 namespace COMETwebapp.Tests.Components.ReferenceData
 {
-    using System;
     using System.Collections.Concurrent;
-    using System.Collections.Generic;
-    using System.Linq;
-    using System.Threading.Tasks;
 
     using Bunit;
 
@@ -45,11 +41,9 @@ namespace COMETwebapp.Tests.Components.ReferenceData
     using COMET.Web.Common.Services.SessionManagement;
     using COMET.Web.Common.Test.Helpers;
 
-    using COMETwebapp.Components.ReferenceData;
+    using COMETwebapp.Components.ReferenceData.Categories;
     using COMETwebapp.Services.ShowHideDeprecatedThingsService;
     using COMETwebapp.ViewModels.Components.ReferenceData.Categories;
-    using COMETwebapp.ViewModels.Components.ReferenceData.Rows;
-    using COMETwebapp.Wrappers;
 
     using DevExpress.Blazor;
 
@@ -318,7 +312,7 @@ namespace COMETwebapp.Tests.Components.ReferenceData
         }
 
         [Test]
-        public async Task VerifyOnInitialized()
+        public async Task VerifyAddingCategory()
         {
             var renderer = this.context.RenderComponent<CategoriesTable>();
 
@@ -330,6 +324,21 @@ namespace COMETwebapp.Tests.Components.ReferenceData
                 Assert.That(renderer.Markup, Does.Contain(this.elementDefinitionCategory1.Name));
                 Assert.That(renderer.Markup, Does.Contain(this.elementDefinitionCategory2.Name));
             });
+
+            this.viewModel.Thing = new Category
+            {
+                Name = "Cat1",
+                ShortName = "TT",
+                IsAbstract = true,
+                IsDeprecated = false
+            };
+
+            this.viewModel.SelectedReferenceDataLibrary = this.siteReferenceDataLibrary;
+            this.viewModel.Thing.PermissibleClass = [ClassKind.ElementDefinition];
+
+            await this.viewModel.CreateCategory(true);
+            this.messageBus.SendMessage(new ObjectChangedEvent(this.viewModel.Thing, EventKind.Added));
+            Assert.That(this.viewModel.Rows.Count, Is.EqualTo(2));
         }
 
         [Test]
@@ -367,6 +376,40 @@ namespace COMETwebapp.Tests.Components.ReferenceData
         }
 
         [Test]
+        public async Task VerifyDisplayCategoryDiagram()
+        {
+            var renderer = this.context.RenderComponent<CategoriesTable>();
+            await renderer.InvokeAsync(() => this.viewModel.SelectCategory(this.elementDefinitionCategory3));
+
+            this.viewModel.CategoryHierarchyDiagramViewModel.SelectedCategory = this.elementDefinitionCategory3;
+            this.viewModel.CategoryHierarchyDiagramViewModel.Rows = this.elementDefinitionCategory3.SuperCategory;
+            this.viewModel.CategoryHierarchyDiagramViewModel.SubCategories = this.elementDefinitionCategory3.SuperCategory;
+
+            await renderer.InvokeAsync(() => this.viewModel.CategoryHierarchyDiagramViewModel.SetupDiagram());
+
+            Assert.Multiple(() =>
+            {
+                Assert.That(this.viewModel.CategoryHierarchyDiagramViewModel.Rows.Count, Is.EqualTo(1));
+                Assert.That(this.viewModel.CategoryHierarchyDiagramViewModel.SubCategories.Count, Is.EqualTo(1));
+            });
+        }
+
+        [Test]
+        public async Task VerifyOnInitialized()
+        {
+            var renderer = this.context.RenderComponent<CategoriesTable>();
+
+            await TaskHelper.WaitWhileAsync(() => this.viewModel.IsLoading);
+
+            Assert.Multiple(() =>
+            {
+                Assert.That(this.viewModel.DataSource.Count, Is.EqualTo(2));
+                Assert.That(renderer.Markup, Does.Contain(this.elementDefinitionCategory1.Name));
+                Assert.That(renderer.Markup, Does.Contain(this.elementDefinitionCategory2.Name));
+            });
+        }
+
+        [Test]
         public async Task VerifyUnDeprecatingCategory()
         {
             var renderer = this.context.RenderComponent<CategoriesTable>();
@@ -400,58 +443,6 @@ namespace COMETwebapp.Tests.Components.ReferenceData
             await this.viewModel.OnConfirmPopupButtonClick();
 
             Assert.That(this.viewModel.IsOnDeprecationMode, Is.False);
-        }
-
-        [Test]
-        public async Task VerifyAddingCategory()
-        {
-            var renderer = this.context.RenderComponent<CategoriesTable>();
-
-            await TaskHelper.WaitWhileAsync(() => this.viewModel.IsLoading);
-
-            Assert.Multiple(() =>
-            {
-                Assert.That(this.viewModel.DataSource.Count, Is.EqualTo(2));
-                Assert.That(renderer.Markup, Does.Contain(this.elementDefinitionCategory1.Name));
-                Assert.That(renderer.Markup, Does.Contain(this.elementDefinitionCategory2.Name));
-            });
-
-            this.viewModel.Thing = new Category
-            {
-                Name = "Cat1",
-                ShortName = "TT",
-                IsAbstract = true,
-                IsDeprecated = false,
-            };
-
-            this.viewModel.SelectedReferenceDataLibrary = this.siteReferenceDataLibrary;
-            this.viewModel.SelectedPermissibleClasses = new List<ClassKindWrapper>() { new (ClassKind.ElementDefinition) };
-
-            await this.viewModel.CreateCategory(TODO);
-            this.messageBus.SendMessage(new ObjectChangedEvent(this.viewModel.Thing, EventKind.Added));
-            Assert.That(this.viewModel.Rows.Count, Is.EqualTo(2));
-        }
-
-        [Test]
-        public async Task VerifyDisplayCategoryDiagram()
-        {
-            var renderer = this.context.RenderComponent<CategoriesTable>();
-
-            await renderer.InvokeAsync(() => this.viewModel.SelectCategory(new CategoryRowViewModel(this.elementDefinitionCategory3)));
-
-            await TaskHelper.WaitWhileAsync(() => this.viewModel.IsLoading);
-
-            this.viewModel.CategoryHierarchyDiagramViewModel.SelectedCategory = this.elementDefinitionCategory3;
-            this.viewModel.CategoryHierarchyDiagramViewModel.Rows = this.elementDefinitionCategory3.SuperCategory;
-            this.viewModel.CategoryHierarchyDiagramViewModel.SubCategories = this.elementDefinitionCategory3.SuperCategory;
-
-            await renderer.InvokeAsync(() => this.viewModel.CategoryHierarchyDiagramViewModel.SetupDiagram());
-
-            Assert.Multiple(() =>
-            {
-                Assert.That(this.viewModel.CategoryHierarchyDiagramViewModel.Rows.Count, Is.EqualTo(1));
-                Assert.That(this.viewModel.CategoryHierarchyDiagramViewModel.SubCategories.Count, Is.EqualTo(1));
-            });
         }
     }
 }
