+++ COMETwebapp/Components/ReferenceData/Categories/CategoriesForm.razor.cs
@@ -1,0 +1,60 @@
+﻿// --------------------------------------------------------------------------------------------------------------------
+//  <copyright file="CategoriesForm.razor.cs" company="Starion Group S.A.">
+//     Copyright (c) 2024 Starion Group S.A.
+// 
+//     Authors: Sam Gerené, Alex Vorobiev, Alexander van Delft, Jaime Bernar, Théate Antoine, João Rua
+// 
+//     This file is part of COMET WEB Community Edition
+//     The COMET WEB Community Edition is the Starion Group Web Application implementation of ECSS-E-TM-10-25 Annex A and Annex C.
+// 
+//     The COMET WEB Community Edition is free software; you can redistribute it and/or
+//     modify it under the terms of the GNU Affero General Public
+//     License as published by the Free Software Foundation; either
+//     version 3 of the License, or (at your option) any later version.
+// 
+//     The COMET WEB Community Edition is distributed in the hope that it will be useful,
+//     but WITHOUT ANY WARRANTY; without even the implied warranty of
+//     MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the GNU
+//    Affero General Public License for more details.
+// 
+//    You should have received a copy of the GNU Affero General Public License
+//    along with this program.  If not, see <http://www.gnu.org/licenses/>.
+//  </copyright>
+//  --------------------------------------------------------------------------------------------------------------------
+
+namespace COMETwebapp.Components.ReferenceData.Categories
+{
+    using System.ComponentModel.DataAnnotations;
+
+    using COMET.Web.Common.Extensions;
+
+    using COMETwebapp.Components.Common;
+    using COMETwebapp.ViewModels.Components.ReferenceData.Categories;
+
+    using Microsoft.AspNetCore.Components;
+
+    using ReactiveUI;
+
+    /// <summary>
+    /// Support class for the <see cref="CategoriesForm" />
+    /// </summary>
+    public partial class CategoriesForm : SelectedDataItemForm
+    {
+        /// <summary>
+        /// The <see cref="ICategoriesTableViewModel" /> for this component
+        /// </summary>
+        [Parameter]
+        [Required]
+        public ICategoriesTableViewModel ViewModel { get; set; }
+
+        /// <summary>
+        /// Method that is executed when there is a valid submit
+        /// </summary>
+        /// <returns>A <see cref="Task" /></returns>
+        protected override async Task OnValidSubmit()
+        {
+            await this.ViewModel.CreateCategory(this.ShouldCreate);
+            await base.OnValidSubmit();
+        }
+    }
+}
