+++ COMETwebapp/Pages/ReferenceData/ReferenceDataPage.razor.cs
@@ -25,6 +25,7 @@
 namespace COMETwebapp.Pages.ReferenceData
 {
     using COMETwebapp.Components.ReferenceData;
+    using COMETwebapp.Components.ReferenceData.Categories;
     using COMETwebapp.Components.ReferenceData.MeasurementScales;
     using COMETwebapp.Components.ReferenceData.ParameterTypes;
